+++ src/table.rs
@@ -1,8 +1,10 @@
 //! Manage BigQuery table
+use std::sync::Arc;
+
 use log::warn;
 use reqwest::Client;
 
-use crate::auth::ServiceAccountAuthenticator;
+use crate::auth::Authenticator;
 use crate::error::BQError;
 use crate::model::get_iam_policy_request::GetIamPolicyRequest;
 use crate::model::policy::Policy;
@@ -17,15 +19,15 @@ use crate::{process_response, urlencode, BIG_QUERY_V2_URL};
 #[derive(Clone)]
 pub struct TableApi {
     client: Client,
-    sa_auth: ServiceAccountAuthenticator,
+    auth: Arc<dyn Authenticator>,
     base_url: String,
 }
 
 impl TableApi {
-    pub(crate) fn new(client: Client, sa_auth: ServiceAccountAuthenticator) -> Self {
+    pub(crate) fn new(client: Client, auth: Arc<dyn Authenticator>) -> Self {
         Self {
             client,
-            sa_auth,
+            auth,
             base_url: BIG_QUERY_V2_URL.to_string(),
         }
     }
@@ -46,7 +48,7 @@ impl TableApi {
             dataset_id = urlencode(&table.table_reference.dataset_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -74,7 +76,7 @@ impl TableApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self.client.delete(req_url.as_str()).bearer_auth(access_token).build()?;
 
@@ -130,7 +132,7 @@ impl TableApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let mut request_builder = self.client.get(req_url.as_str()).bearer_auth(access_token);
         if let Some(selected_fields) = selected_fields {
@@ -158,7 +160,7 @@ impl TableApi {
             dataset_id = urlencode(dataset_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let mut request = self.client.get(req_url).bearer_auth(access_token);
 
@@ -199,7 +201,7 @@ impl TableApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -234,7 +236,7 @@ impl TableApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -263,7 +265,7 @@ impl TableApi {
             resource = urlencode(resource)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -292,7 +294,7 @@ impl TableApi {
             resource = urlencode(resource)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -324,7 +326,7 @@ impl TableApi {
             resource = urlencode(resource)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client