+++ src/tabledata.rs
@@ -1,5 +1,7 @@
 //! Manage BigQuery streaming API.
-use crate::auth::ServiceAccountAuthenticator;
+use std::sync::Arc;
+
+use crate::auth::Authenticator;
 use crate::error::BQError;
 use crate::model::data_format_options::DataFormatOptions;
 use crate::model::table_data_insert_all_request::TableDataInsertAllRequest;
@@ -11,15 +13,15 @@ use reqwest::Client;
 #[derive(Clone)]
 pub struct TableDataApi {
     client: Client,
-    sa_auth: ServiceAccountAuthenticator,
+    auth: Arc<dyn Authenticator>,
     base_url: String,
 }
 
 impl TableDataApi {
-    pub(crate) fn new(client: Client, sa_auth: ServiceAccountAuthenticator) -> Self {
+    pub(crate) fn new(client: Client, auth: Arc<dyn Authenticator>) -> Self {
         Self {
             client,
-            sa_auth,
+            auth,
             base_url: BIG_QUERY_V2_URL.to_string(),
         }
     }
@@ -51,7 +53,7 @@ impl TableDataApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client
@@ -86,7 +88,7 @@ impl TableDataApi {
             table_id = urlencode(table_id)
         );
 
-        let access_token = self.sa_auth.access_token().await?;
+        let access_token = self.auth.access_token().await?;
 
         let request = self
             .client