+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.15.0] - 2022-11-02
+
+### Improvement
+
+- Abstraction of the Authenticator (Thanks @kiibo382).
+- InstalledFlowAuthenticator support (Thanks @kiibo382).
+- Bump yup-oauth2 crate version.
+- Bump time crate version (security issue).
+
 ## [0.14.0] - 2022-09-27
 
 ### Fix